+++ lib/build-manager/component-provider/index.js
@@ -2,8 +2,8 @@
 
 const _ = require('nami-utils/lodash-extra');
 const Logger = require('nami-logger');
-const Recipe = require('./recipe');
-
+const RecipeLogicProvider = require('./recipe/logic-provider');
+const RecipeMetadataProvider = require('./recipe/metadata-provider');
 
 /**
  * Class representing the Component Provider
@@ -35,7 +35,12 @@ class ComponentProvider {
       };
     }
     this.recipeDirectories = _.toArrayIfNeeded(recipeDirectories);
-    this.componentTypeCollections = componentTypeCollections;
+    this.metadataProvider = new RecipeMetadataProvider(this.recipeDirectories, options.metadataServer, {
+      logger: options.logger
+    });
+    this.logicProvider = new RecipeLogicProvider(this.recipeDirectories, componentTypeCollections, {
+      logger: options.logger
+    });
     this.logger = options.logger || new Logger({level: 'silent'});
   }
 
@@ -57,11 +62,11 @@ class ComponentProvider {
    * getRecipe('zlib', {version: '~1.2'});
    */
   getRecipe(id, requirements) {
-    return new Recipe(id, this.recipeDirectories, this.componentTypeCollections, {
-      requirements,
-      logger: this.logger,
-      metadataServer: this.metadataServer
-    });
+    const recipeDir = this.logicProvider.findRecipeFolder(id, this.recipeDirectories);
+    return {
+      metadata: this.metadataProvider.getMetadata(id, {recipeDir, requirements}),
+      componentClass: this.logicProvider.getRecipeClass(id, recipeDir, requirements)
+    };
   }
 
   /**
