+++ lib/build-manager/component-provider/recipe/spec.js
